adds javaDoc for laser grid system and comp, makes laser grid passable for objects
